requirements.txt
canvas-lms-api==0.4.1
canvasapi==2.0.0
dateparser==0.7.6
freezegun==1.0.0
python-dateutil==2.8.1
spacy
tzlocal
urllib3
pytz
redis
thinc
pynacl
en-core-web-sm
toga
briefcase
regex==2022.3.2